+++ tests/resolvers/Query/user.spec.ts
@@ -6,7 +6,7 @@ import { User } from "../../../src/models";
 import { user as userResolver } from "../../../src/resolvers/Query/user";
 import { connect, disconnect } from "../../helpers/db";
 
-import { afterAll, beforeAll, describe, expect, it, vi } from "vitest";
+import { afterAll, beforeAll, describe, expect, it } from "vitest";
 import { deleteUserFromCache } from "../../../src/services/UserCache/deleteUserFromCache";
 import type { QueryUserArgs } from "../../../src/types/generatedGraphQLTypes";
 import type { TestUserType } from "../../helpers/userAndOrg";
@@ -63,56 +63,5 @@ describe("resolvers -> Query -> user", () => {
       organizationsBlockedBy: [],
       image: null,
     });
-  });
-
-  it("throws NotFoundError when the specified user ID does not exist", async () => {
-    expect.assertions(2);
-    const nonExistentUserId = new Types.ObjectId().toString();
-    const args = {
-      id: nonExistentUserId,
-    };
-
-    const context = {
-      userId: new Types.ObjectId().toString(),
-    };
-
-    if (typeof userResolver === "function") {
-      try {
-        await userResolver({}, args, context);
-      } catch (error: unknown) {
-        expect(error).toBeInstanceOf(Error);
-        expect((error as Error).message).toEqual(USER_NOT_FOUND_ERROR.DESC);
-      }
-    } else {
-      throw new Error("userResolver is not defined");
-    }
-  });
-
-  it("throws NotFoundError when fetching user profile and user is null", async () => {
-    expect.assertions(2);
-    const args = {
-      id: new Types.ObjectId().toString(),
-    };
-
-    const context = {
-      userId: new Types.ObjectId().toString(),
-    };
-
-    // Mock User.findById to return null
-    vi.spyOn(User, "findById").mockResolvedValueOnce(null);
-
-    if (typeof userResolver === "function") {
-      try {
-        await userResolver({}, args, context);
-      } catch (error: unknown) {
-        expect(error).toBeInstanceOf(Error);
-        expect((error as Error).message).toEqual(USER_NOT_FOUND_ERROR.DESC);
-      }
-    } else {
-      throw new Error("userResolver is not defined");
-    }
-
-    // Restore original implementation
-    vi.restoreAllMocks();
   });
 });
